+++ daml-lf/data/src/main/scala/com/digitalasset/daml/lf/data/InsertOrdSet.scala
@@ -3,6 +3,8 @@
 
 package com.daml.lf.data
 
+import scala.collection.immutable.{HashSet, Queue}
+
 /**
   * Insert-ordered Set.
   *
@@ -12,21 +14,8 @@ package com.daml.lf.data
   *  insert: O(1)
   *  remove: O(n)
   */
-import scala.collection.immutable.{HashSet, Set, Queue}
-import scala.collection.{SetLike, AbstractSet}
-import scala.collection.generic.{
-  ImmutableSetFactory,
-  GenericCompanion,
-  CanBuildFrom,
-  GenericSetTemplate
-}
-
 final class InsertOrdSet[T] private (_items: Queue[T], _hashSet: HashSet[T])
-    extends AbstractSet[T]
-    with Set[T]
-    with SetLike[T, InsertOrdSet[T]]
-    with GenericSetTemplate[T, InsertOrdSet]
-    with Serializable {
+    extends AbstractInsertOrdSet[T] {
   override def empty: InsertOrdSet[T] = InsertOrdSet.empty
   override def size: Int = _hashSet.size
 
@@ -36,7 +25,7 @@ final class InsertOrdSet[T] private (_items: Queue[T], _hashSet: HashSet[T])
   override def contains(elem: T): Boolean =
     _hashSet.contains(elem)
 
-  override def +(elem: T): InsertOrdSet[T] =
+  override def incl(elem: T): InsertOrdSet[T] =
     if (_hashSet.contains(elem))
       this
     else
@@ -45,25 +34,23 @@ final class InsertOrdSet[T] private (_items: Queue[T], _hashSet: HashSet[T])
         _hashSet + elem
       )
 
-  override def -(elem: T): InsertOrdSet[T] =
+  override def excl(elem: T): InsertOrdSet[T] =
     new InsertOrdSet(
       _items.filter(elem2 => elem != elem2),
       _hashSet - elem
     )
-
-  override def companion: GenericCompanion[InsertOrdSet] = InsertOrdSet
-
 }
 
-object InsertOrdSet extends ImmutableSetFactory[InsertOrdSet] {
+object InsertOrdSet extends InsertOrdSetCompanion {
   private val Empty = new InsertOrdSet(Queue.empty, HashSet.empty)
   override def empty[T] = Empty.asInstanceOf[InsertOrdSet[T]]
-  def emptyInstance: InsertOrdSet[Any] = empty[Any]
 
   def fromSeq[T](s: Seq[T]): InsertOrdSet[T] =
     new InsertOrdSet(Queue(s.reverse: _*), HashSet(s: _*))
 
-  implicit def canBuildFrom[A]: CanBuildFrom[Coll, A, InsertOrdSet[A]] =
-    setCanBuildFrom[A]
+  // Here only for 2.12 (harmless in 2.13); placed in InsertOrdSetCompanion the
+  // implicit gets in an unwinnable fight with Set's version
+  implicit override def canBuildFrom[A]: Factory[A] =
+    super.canBuildFrom
 
 }
